Moved reusable config classes to their own namespace
