+++ src/Config/ConfigAwareTrait.php
@@ -5,7 +5,7 @@
  * @package MultilingualPress2to3
  */
 
-namespace Inpsyde\MultilingualPress2to3;
+namespace Inpsyde\MultilingualPress2to3\Config;
 
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
