+++ src/IntegrationHandler.php
@@ -3,6 +3,7 @@
 namespace Inpsyde\MultilingualPress2to3;
 
 use Dhii\I18n\StringTranslatingTrait;
+use Inpsyde\MultilingualPress2to3\Config\ConfigAwareTrait;
 use Inpsyde\MultilingualPress2to3\Handler\HandlerInterface;
 use Inpsyde\MultilingualPress2to3\Handler\HandlerTrait;
 use Psr\Container\ContainerInterface;
